requirements.txt
# Full requirements for CMS pipeline container

# Azure Blob Storage  
azure-storage-blob>=12.19.0

# Data Processing
pandas==2.2.0
numpy==1.26.0

# Machine Learning
scikit-learn==1.4.0

# HTTP Requests
requests>=2.31.0

# Other Dependencies
python-dateutil>=2.8.2
